+++ src/rainfallqc/qc_frameworks/inbuilt_qc_frameworks.py
@@ -1,13 +1,20 @@
 # -*- coding: utf-8 -*-
 """In-built QC frameworks to apply to rainfall data to create quality controlled data."""
 
-from rainfallqc.checks import comparison_checks, gauge_checks, neighbourhood_checks, pypwsqc_filters, timeseries_checks
+from rainfallqc.checks import (
+    comparison_checks,
+    gauge_checks,
+    neighbourhood_checks,
+    pypwsqc_filters,
+    subhourlyqc_checks,
+    timeseries_checks,
+)
 
 INTENSE_QC = {
     "QC1": {"function": gauge_checks.check_years_where_nth_percentile_is_zero},
     "QC2": {"function": gauge_checks.check_years_where_annual_kth_largest_value_is_zero},
-    "QC3": {"function": gauge_checks.check_temporal_bias},
-    "QC4": {"function": gauge_checks.check_temporal_bias},
+    "QC3": {"function": gauge_checks.check_day_of_week},
+    "QC4": {"function": gauge_checks.check_hour_of_day},
     "QC5": {"function": gauge_checks.check_intermittency},
     "QC6": {"function": gauge_checks.check_breakpoints},
     "QC7": {"function": gauge_checks.check_min_val_change},
@@ -64,6 +71,33 @@ INTENSE_RULEBASE_QC = {
     },
 }
 
+# SubHourlyQC from Villalobos-Herrera et al. 2022 (https://doi.org/10.1002/qj.4357). Also see Table S1 of that paper.
+UK_SUBHOURLY_QC = {
+    # From IntenseQC (12)
+    "HQC_QC13": {"function": timeseries_checks.check_daily_accumulations},
+    "HQC_QC14": {"function": timeseries_checks.check_monthly_accumulations},
+    "HQC_QC5": {"function": gauge_checks.check_intermittency},
+    "HQC_QC1": {"function": gauge_checks.check_years_where_nth_percentile_is_zero},
+    "HQC_QC2": {"function": gauge_checks.check_years_where_annual_kth_largest_value_is_zero},
+    "HQC_QC3": {"function": gauge_checks.check_day_of_week},
+    "HQC_QC4": {"function": gauge_checks.check_hour_of_day},
+    "HQC_QC6": {"function": gauge_checks.check_breakpoints},
+    "HQC_QC17": {"function": neighbourhood_checks.check_wet_neighbours_hourly},
+    "HQC_QC19": {"function": neighbourhood_checks.check_dry_neighbours_hourly},
+    "HQC_QC8": {"function": comparison_checks.check_annual_exceedance_etccdi_r99p},
+    "HQC_QC9": {"function": comparison_checks.check_annual_exceedance_etccdi_prcptot},
+    # Modified part of IntenseQC (4)
+    "HQC_UK1hr": {"function": subhourlyqc_checks.check_exceedance_of_UK_1hr_record},
+    "HQC_UK24hr": {"function": subhourlyqc_checks.check_exceedance_of_UK_24hr_record},
+    "HQC_UK24hr_rolling": {"function": subhourlyqc_checks.check_daily_exceedance_of_UK_24hr_record},
+    "HQC_streaks_20mm": {"function": subhourlyqc_checks.check_streaks_20mm},
+    # New checks (3)
+    # "SHQC_QC_spike_check": {"function": subhourlyqc_checks.spike_check}, # awaiting chat with Roberto
+    "SHQC_freqResChecker": {"function": subhourlyqc_checks.check_freq_is_subhourly},
+    "SHQC_subH_checkr": {"function": subhourlyqc_checks.check_subhourly_thresholds},
+}
+
+
 PYPWSQC = {
     "BC": {"function": pypwsqc_filters.run_bias_correction},
     "EBF": {"function": pypwsqc_filters.run_event_based_filter},
@@ -74,7 +108,12 @@ PYPWSQC = {
     "SO": {"function": pypwsqc_filters.check_station_outlier},
 }
 
-INBUILT_QC_FRAMEWORKS = {"intenseqc": INTENSE_QC, "intenseqc_rulebase_only": INTENSE_RULEBASE_QC, "pypwsqc": PYPWSQC}
+INBUILT_QC_FRAMEWORKS = {
+    "intenseqc": INTENSE_QC,
+    "intenseqc_rulebase_only": INTENSE_RULEBASE_QC,
+    "uk_subhourlyqc": UK_SUBHOURLY_QC,
+    "pypwsqc": PYPWSQC,
+}
 
 
 # all checks that are computed as summary of overall data or once per year values