+++ tests/unit/test_gauge_checks.py
@@ -36,23 +36,23 @@ def test_get_years_where_annual_mean_k_top_rows_are_zero(hourly_gsdr_data):
 
 
 def test_check_day_of_week_bias(daily_gsdr_data):
-    week_bias_true = gauge_checks.check_temporal_bias(
-        daily_gsdr_data, target_gauge_col=DEFAULT_RAIN_COL, time_granularity="weekday"
+    week_bias_true = gauge_checks.check_day_of_week(
+        daily_gsdr_data, target_gauge_col=DEFAULT_RAIN_COL,
     )
     assert week_bias_true == 1
     # week_bias_false = gauge_checks.check_day_of_week_bias(other_daily_gsdr_data, target_gauge_col=DEFAULT_RAIN_COL)
 
 
 def test_check_hour_of_day_bias(hourly_gsdr_data):
-    hour_bias_true = gauge_checks.check_temporal_bias(
-        hourly_gsdr_data, target_gauge_col=DEFAULT_RAIN_COL, time_granularity="hour"
+    hour_bias_true = gauge_checks.check_hour_of_day(
+        hourly_gsdr_data, target_gauge_col=DEFAULT_RAIN_COL,
     )
     assert hour_bias_true == 0
 
 
 def test_check_wrong_time_gran(daily_gsdr_data):
     with pytest.raises(ValueError):
-        gauge_checks.check_temporal_bias(daily_gsdr_data, target_gauge_col=DEFAULT_RAIN_COL, time_granularity="wrong")
+        gauge_checks.temporal_bias_test(daily_gsdr_data, target_gauge_col=DEFAULT_RAIN_COL, time_granularity="wrong")
 
 
 def test_intermittency_check(gappy_daily_data):