+++ tests/unit/test_neighbourhood_checks.py
@@ -265,7 +265,7 @@ def test_check_monthly_neighbours_15min(mins15_gsdr_network):
         time_res="15m",
     )
 
-    assert len(result.filter(pl.col("majority_monthly_flag") == 1)) == 720*4
+    assert len(result.filter(pl.col("majority_monthly_flag") == 1)) == 2880
     assert len(result.filter(pl.col("majority_monthly_flag") == 3)) == 2976
 
 
@@ -436,7 +436,7 @@ def test_check_monthly_factor(monthly_gsdr_network, monthly_gpcc_network):
         nearest_neighbour=f"{DEFAULT_RAIN_COL}_DE_00310",
     )
     assert round(result["monthly_factor_flag"].max(), 2) == 6
-    assert len(result.filter(pl.col("monthly_factor_flag") > 0)) == 6
+    assert len(result.filter(pl.col("monthly_factor_flag").fill_nan(0.0) > 0)) == 6
 
     result = neighbourhood_checks.check_monthly_factor(
         monthly_gpcc_network,
@@ -445,7 +445,7 @@ def test_check_monthly_factor(monthly_gsdr_network, monthly_gpcc_network):
     )
 
     assert round(result["monthly_factor_flag"].max(), 2) == 3
-    assert len(result.filter(pl.col("monthly_factor_flag") > 0)) == 61
+    assert len(result.filter(pl.col("monthly_factor_flag").fill_nan(0.0) > 0)) == 61
 
 
 def test_make_num_neighbours_online_col(hourly_gsdr_network):